+++ alice-facade-tui/src/main/java/module-info.java
@@ -14,6 +14,7 @@ module alice.agent.facade.tui.main {
   requires alice.agent.alice.core.agent.main;
   requires alice.agent.alice.env.adapter.main;
   requires alice.agent.alice.model.main;
+  requires alice.agent.alice.tool.gateway.main;
   requires alice.agent.command.main;
   requires alice.agent.alice.memory.vault.main;